trainee_score table, scores: int => dec(3,2)

+++ db/sql/02-create-table-trainee-score.sql
@@ -3,10 +3,10 @@ DROP TABLE IF EXISTS trainee_score;
 CREATE TABLE trainee_score
 (
 	id                VARCHAR(36) PRIMARY KEY DEFAULT (UUID()),
-	courseCompletion  TINYINT NOT NULL,
-	courseEngagment   TINYINT NOT NULL,
-	projectDegree     TINYINT NOT NULL,
-	teamProjectDegree TINYINT NOT NULL,
+	courseCompletion  DECIMAL(3,2) NOT NULL,
+	courseEngagment   DECIMAL(3,2) NOT NULL,
+	projectDegree     DECIMAL(3,2) NOT NULL,
+	teamProjectDegree DECIMAL(3,2) NOT NULL,
 	bonusProjectUrls  JSON    NOT NULL,
 	userId            VARCHAR(36),
 	createdAt         DATE    NOT NULL        DEFAULT (NOW()),
